extract duplicate methods to traits
