+++ api/GitHub/Hook.php
@@ -3,6 +3,9 @@
 namespace GitHub;
 
 use \EventMap;
+use \Traits\MapsEvents;
+use \Traits\HandlesEvents;
+use \Traits\Authenticates;
 
 /**
  * GitHub service for the webhook-api.
@@ -17,26 +20,8 @@ use \EventMap;
  */
 class Hook extends \Hook
 {
-    /**
-     * Authorization key to be provided from the user.
-     *
-     * @var string null
-     */
-    private $secret = null;
 
-    /**
-     * The X-GitHub-Event header, i.e push, issue, etc.
-     *
-     * @var string
-     */
-    private $event;
-
-    /**
-     * The hash string from the X-GitHub-Signature header.
-     *
-     * @var string
-     */
-    private $signature;
+    use MapsEvents, HandlesEvents, Authenticates;
 
     /**
      * The hash algorithm used in the X-GitHub-Signature header.
@@ -44,27 +29,6 @@ class Hook extends \Hook
      * @var string
      */
     private $algorithm;
-
-    /**
-     * The events the API will listen to.
-     *
-     * @var array
-     */
-    public $listeners = [];
-
-    /**
-     * The default available events to listen to.
-     *
-     * @var array
-     */
-    private $defaultListeners;
-
-    /**
-     * The GitHub Event map
-     *
-     * @var array
-     */
-    private $eventMap;
 
     /**
      * Checking for X-GitHub-Event header and authorizing.
@@ -94,23 +58,12 @@ class Hook extends \Hook
     }
 
     /**
-     * Set the Event Map and the default listeners.
-     *
-     * @param array $map
-     */
-    private function setEventMap($map)
-    {
-        $this->eventMap = $map;
-        $this->defaultListeners = array_keys($map);
-    }
-
-    /**
      * Authorizing method with the helper functions secretValidator() and checkSecret()
      * Sends message to apiMessages if a problem occurs.
      *
      * @return bool true | false
      */
-    protected function auth()
+    private function auth()
     {
         if (!array_key_exists('X-Hub-Signature', $this->headers)) {
             $this->apiMessages[] = 'No signature provided';
@@ -136,97 +89,14 @@ class Hook extends \Hook
      *
      * @return bool hash
      */
-    protected function checkSecret()
+    private function checkSecret()
     {
         return hash_equals(
             hash_hmac(
                 $this->algorithm,
                 $this->secretValidator(),
                 $this->secret
-            ),  $this->signature);
-    }
-
-    /**
-     * Returns the payload temporarily for authorization needed in checkSecret().
-     *
-     * @return array payload
-     */
-    protected function secretValidator()
-    {
-        switch ($this->contentType) {
-            case 'application/json':
-                return file_get_contents('php://input');
-            break;
-            case 'application/x-www-form-urlencoded':
-                return \Request::input('payload');
-            break;
-        }
-    }
-
-    /**
-     * Sets the events to listen to
-     * Needs to be declared to create any output
-     * Empty array watches all events
-     * Can be complemented with callback functions.
-     *
-     * @param array $listeners Array of events
-     *
-     * @return object | false   false if event is not being watched
-     */
-    public function listen(array $listeners)
-    {
-        if (empty($listeners)) {
-            $listeners = $this->defaultListeners;
-        }
-
-        foreach ($listeners as $listener => $callback) {
-            if (!in_array($listener, $this->defaultListeners) && !in_array($callback, $this->defaultListeners)) {
-                $this->apiMessages[] = "Can't watch an invalid event";
-
-                return false;
-            }
-        }
-
-        $this->listeners = $listeners;
-
-        if ($this->notWatchingEvent()) {
-            return false;
-        }
-
-        $this->registerEvent();
-
-        if (isset($this->listeners[$this->event])) {
-            call_user_func($this->listeners[$this->event], $this->output->__toString());
-        }
-    }
-
-    /**
-     * Check if the event is set up to be watched.
-     *
-     * @return bool
-     */
-    private function notWatchingEvent()
-    {
-        if (!array_key_exists($this->event, $this->listeners) && !in_array($this->event, $this->listeners)) {
-            if (!empty($this->event)) {
-                $this->apiMessages[] = "Not watching $this->event event";
-            }
-
-            return false;
-        }
-
-        return true;
-    }
-
-    /**
-     * Register the event and send the payload to its configured class.
-     *
-     * @return void
-     */
-    private function registerEvent()
-    {
-        if (array_key_exists($this->event, $this->eventMap)) {
-            $this->output = new $this->eventMap[$this->event]($this->payload);
-        }
+            ),
+            $this->signature);
     }
 }
